feat: added query and mutations for package to backend
